requirements.txt
python-telegram-bot==21.9
telethon==1.36.0
playwright==1.49.1
playwright-stealth==1.0.6
motor==3.6.0
ffmpeg-python==0.2.0
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
aria2p==0.12.0
requests==2.32.3
beautifulsoup4==4.12.3
